requirements.txt
appnope
asttokens
bleach==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
comm
contourpy==1.3.2
cycler==0.12.1
debugpy
decorator
exceptiongroup
executing
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.3.2
idna==3.10
importlib_metadata
ipykernel
ipython
ipython_pygments_lexers
jedi
Jinja2==3.1.6
joblib==1.5.0
jupyter_client
jupyter_core
kaggle==1.7.4.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline
mpmath==1.3.0
nest_asyncio
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pandas==2.2.3
parso
pexpect
pickleshare
pillow==11.2.1
platformdirs
prompt_toolkit
protobuf==6.31.0
psutil
ptyprocess
pure_eval
Pygments
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
pyzmq
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
stack_data
sympy==1.14.0
text-unidecode==1.3
threadpoolctl==3.6.0
torch==2.7.0
torchvision==0.22.0
tornado
tqdm==4.67.1
traitlets
typing_extensions
tzdata==2025.2
urllib3==2.4.0
wcwidth
webencodings==0.5.1
zipp
